refactor: major codebase cleanup and Price Rule Wizard
## Frontend Refactoring
- Migrate FormData and API types to snake_case (align with Rust backend)
- Add LabelPrintState type (-1=inherit, 0=disabled, 1=enabled)
- Rename OrderStatus to OrderApiStatus (avoid conflict with event sourcing)
- Update i18n keys to snake_case convention
- Delete duplicate utility files (formatCurrency, checkoutCalculations, attributeDisplay, product.ts)
- Delete misplaced TypeScript files from src-tauri/src/infrastructure/

## Price Rule Wizard (New Feature)
- Add 6-step wizard for creating price rules
- Step1: Rule type selection (discount/surcharge)
- Step2: Adjustment configuration (percentage/fixed)
- Step3: Scope selection (global/category/tag/product)
- Step4: Time configuration (always/schedule/one-time)
- Step5: Naming (name, display_name, receipt_name)
- Step6: Advanced options (priority, stackable, exclusive)

## Backend Changes
- Update crab-client login API to return connection state on failure
- Update edge-server models for consistency

+++ edge-server/src/db/models/attribute.rs
@@ -13,9 +13,9 @@ pub type AttributeId = Thing;
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct AttributeOption {
     pub name: String,
-    /// Price modifier in cents (positive=add, negative=subtract)
+    /// Price modifier in currency unit (positive=add, negative=subtract, e.g., 2.50 = ¥2.50)
     #[serde(default)]
-    pub price_modifier: i64,
+    pub price_modifier: f64,
     #[serde(default)]
     pub display_order: i32,
     #[serde(
@@ -35,7 +35,7 @@ impl AttributeOption {
     pub fn new(name: String) -> Self {
         Self {
             name,
-            price_modifier: 0,
+            price_modifier: 0.0,
             display_order: 0,
             is_active: true,
             receipt_name: None,
